HUGE optimization TEST THE GAME PLEASE
CODE:        7783 → 7580 - diff:  203
LVL_BANK_00: 3693 → 3680 - diff:  13
RODATA:      341  → 322  - diff:  19
XCD_BANK_00: 4396 → 4108 - diff:  288
XCD_BANK_01: 6117 → 5784 - diff:  333
XCD_BANK_04: 7965 → 7834 - diff:  131

Total size reduction: 987 bytes
Fixed bank space reduction: 222 bytes

+++ LEVELS/levellist.h
@@ -67,7 +67,7 @@ const uint8_t difficulty_pal_B[] ={
 	0x21,	// danger
 };
 
-const unsigned short DECOTYPE[] = {
+const uint8_t DECOTYPE[] = {
 	DECO1,
 	DECO1,
 	DECO1,
